backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
openai>=1.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium>=4.15.0
Pillow>=10.0.0
celery>=5.3.0
redis>=5.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
pandas>=2.0.0
imbalanced-learn>=0.11.0
joblib>=1.3.0
psycopg2-binary>=2.9.0
django-cache-machine>=1.2.0

